Update Subsequence Hate.cpp

+++ 1400/Subsequence Hate.cpp
@@ -12,7 +12,7 @@ int main()
 		string s;
 		cin>>s;
 		ll ones=0;
-		int zeroes=0;
+		ll zeroes=0;
 		for(ll i=0 ;i<s.length(); i++)
 		{
 			if(s[i]=='0')
